backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.10
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
Pillow==10.4.0
httpx==0.27.2
